added ARM ADD #7 instruction
Former-commit-id: 94439cd064471dfaada3a5037438c19a08f85526

+++ src/jumptable/jumptable-thumb-config.cpp
@@ -607,7 +607,7 @@ void run_1010SREG(uint16_t opcode) {
 
 // add / subtract offset to stack pointer
 void run_10110000(uint16_t opcode) {
-    uint8_t offset      = get_nth_bits(opcode, 0, 7) << 2;
+    uint16_t offset     = get_nth_bits(opcode, 0, 7) << 2;
     bool is_subtraction = get_nth_bit(opcode, 7);
     
     if (is_subtraction) {
